+++ server/build.rs
@@ -35,6 +35,7 @@ fn main() {
 
     for proto in &[
         "../proto/vstimd/v1/vec2.proto",
+        "../proto/vstimd/v1/transform.proto",
         "../proto/vstimd/v1/color.proto",
         "../proto/vstimd/v1/stimuli/stimulus_type.proto",
         "../proto/vstimd/v1/stimuli/shapes.proto",
@@ -60,6 +61,7 @@ fn main() {
         .compile_protos(
             &[
                 "../proto/vstimd/v1/vec2.proto",
+                "../proto/vstimd/v1/transform.proto",
                 "../proto/vstimd/v1/color.proto",
                 "../proto/vstimd/v1/stimuli/stimulus_type.proto",
                 "../proto/vstimd/v1/stimuli/shapes.proto",